Add ok property to Response
This is intended to replicate the ok property on the requests Response
object.  When porting existing requests code to httpx, it would be
useful if this path existed, and had similar behavior to the requests
library, to aid in porting code from one to another.

The requests library implements this as a call to raise_for_status(),
while catching the exception[2].  This patchset opts for similar
behavior.

[1] https://requests.readthedocs.io/en/latest/api/#requests.Response.ok
[2] https://github.com/psf/requests/blob/f1bb07d39b74d6444e333879f8b8a3d9dd4d2311/src/requests/models.py#L755

Signed-off-by: Ed Tanous <ed@tanous.net>